+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Cancel all tasks when cancelling a compute plan (#546)
 - Check for compute plan existence on task registration (#554)
+- Disallow registration of tasks on a compute plan you don't own (#566)
 
 ## [0.6.0] - 2022-02-18